+++ web-login-RayMond/src/main/java/com/example/demo/service/UserService.java
@@ -22,7 +22,7 @@ public class UserService {
 	}
 	
 	
-	  public Optional<NewUser> findById(Long id) {
+	  public Optional<NewUser> findById(int id) {
 	        return userRepository.findById(id);
 	    }
 
@@ -31,7 +31,7 @@ public class UserService {
 	}
 
 
-	public void delete(Long id) {
+	public void delete(int id) {
 		userRepository.deleteById(id);
 		
 	}
